# openapi/personalcrm_openapi_v1.yaml
openapi: 3.0.3
info:
  title: PersonalCRM API
  version: 1.0.0
  description: |
    PersonalCRM v1 API for managing contacts, interactions, introductions, reminders,
    strategy/wheel, projects, and RBAC in a cloud workspace with assistant access.

    Notes:
    - All requests are scoped to a workspace.
    - Field-level privacy is enforced server-side; assistants may receive redacted fields.
    - Use cursor-based pagination (next_cursor).
servers:
  - url: https://api.personalcrm.example.com
    description: Production (replace)
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/rbac/roles:
    get:
      tags: [RBAC]
      summary: List workspace roles
      operationId: listRoles
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdQuery'
      responses:
        '200':
          description: Roles
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Role'
                required: [data]
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [RBAC]
      summary: Create a role
      operationId: createRole
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdQuery'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/v1/rbac/permissions:
    get:
      tags: [RBAC]
      summary: List known permissions
      operationId: listPermissions
      responses:
        '200':
          description: Permissions
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Permission'
                required: [data]
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/contacts:
    get:
      tags: [Contacts]
      summary: List contacts
      operationId: listContacts
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdQuery'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - name: q
          in: query
          schema:
            type: string
          description: Full-text search across name, org, title, shared notes
        - name: tags
          in: query
          schema:
            type: array
            items: { type: string }
          style: form
          explode: true
          description: Filter by tags (OR semantics)
        - name: tier
          in: query
          schema:
            $ref: '#/components/schemas/TieStrength'
          description: Filter by tie strength (close/medium/weak)
        - name: visibility
          in: query
          schema:
            $ref: '#/components/schemas/Visibility'
          description: Filter by record visibility
        - name: updated_since
          in: query
          schema:
            type: string
            format: date-time
          description: Return contacts updated after this timestamp
        - name: sort
          in: query
          schema:
            type: string
            enum: [updated_desc, last_interaction_desc, name_asc]
          description: Sort option
      responses:
        '200':
          description: Contacts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedContacts'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Contacts]
      summary: Create a contact
      operationId: createContact
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdQuery'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContactCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/contacts/{contact_id}:
    get:
      tags: [Contacts]
      summary: Get a contact
      operationId: getContact
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdQuery'
        - $ref: '#/components/parameters/ContactIdPath'
      responses:
        '200':
          description: Contact
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Contacts]
      summary: Update a contact
      operationId: updateContact
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdQuery'
        - $ref: '#/components/parameters/ContactIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContactUpdate'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Contacts]
      summary: Delete a contact
      operationId: deleteContact
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdQuery'
        - $ref: '#/components/parameters/ContactIdPath'
      responses:
        '204':
          description: Deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/contacts/{contact_id}/interactions:
    get:
      tags: [Interactions]
      summary: List interactions for a contact
      operationId: listContactInteractions
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdQuery'
        - $ref: '#/components/parameters/ContactIdPath'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/InteractionType'
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Interactions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedInteractions'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      tags: [Interactions]
      summary: Create an interaction for a contact
      operationId: createContactInteraction
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdQuery'
        - $ref: '#/components/parameters/ContactIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InteractionCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Interaction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/reminders:
    get:
      tags: [Reminders]
      summary: List reminders
      operationId: listReminders
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdQuery'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ReminderStatus'
        - name: due_from
          in: query
          schema:
            type: string
            format: date-time
        - name: due_to
          in: query
          schema:
            type: string
            format: date-time
        - name: contact_id
          in: query
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: Reminders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedReminders'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Reminders]
      summary: Create a reminder
      operationId: createReminder
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdQuery'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReminderCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reminder'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/reminders/{reminder_id}:
    patch:
      tags: [Reminders]
      summary: Update a reminder
      operationId: updateReminder
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdQuery'
        - $ref: '#/components/parameters/ReminderIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReminderUpdate'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reminder'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/projects:
    get:
      tags: [Projects]
      summary: List projects
      operationId: listProjects
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdQuery'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - name: q
          in: query
          schema:
            type: string
          description: Search by name/description
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ProjectStatus'
      responses:
        '200':
          description: Projects
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedProjects'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Projects]
      summary: Create a project
      operationId: createProject
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdQuery'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    WorkspaceIdQuery:
      name: workspace_id
      in: query
      required: true
      schema:
        $ref: '#/components/schemas/UUID'
      description: Workspace scope for the request
...
    ContactIdPath:
      name: contact_id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/UUID'
...
    ReminderIdPath:
      name: reminder_id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/UUID'
...
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
      description: Opaque pagination cursor
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    UUID:
      type: string
      format: uuid
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
          required: [code, message]
      required: [error]
    Visibility:
      type: string
      enum: [private, shared, limited]
      description: |
        Record visibility.
        - private: only Owner (and roles with private access)
        - shared: Owner + Assistant
        - limited: Assistant can see a redacted subset
    TieStrength:
      type: string
      enum: [close, medium, weak]
    InteractionType:
      type: string
      enum: [meeting, call, message, event, intro, help_given, help_received, note]
...
    ReminderStatus:
      type: string
      enum: [open, done, canceled]
    ProjectStatus:
      type: string
      enum: [idea, active, on_hold, completed, canceled]
...
    Permission:
      type: object
      properties:
        key: { type: string }
        description: { type: string }
      required: [key]
    Role:
      type: object
      properties:
        id: { $ref: '#/components/schemas/UUID' }
        workspace_id: { $ref: '#/components/schemas/UUID' }
        name: { type: string }
        description: { type: string }
        permission_keys:
          type: array
          items: { type: string }
      required: [id, workspace_id, name, permission_keys]
    RoleCreate:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        permission_keys:
          type: array
          items: { type: string }
      required: [name, permission_keys]
...
    Contact:
      type: object
      properties:
        id: { $ref: '#/components/schemas/UUID' }
        workspace_id: { $ref: '#/components/schemas/UUID' }
        visibility: { $ref: '#/components/schemas/Visibility' }

        first_name: { type: string }
        last_name: { type: string }
        middle_name: { type: string }
        display_name: { type: string }
        photo_url: { type: string }

        emails:
          type: array
          items: { type: string, format: email }
        phones:
          type: array
          items: { type: string }
        messengers:
          type: object
          additionalProperties: { type: string }
          description: Map of messenger name -> handle

        city: { type: string }
        timezone: { type: string }
        birthday: { type: string, format: date }

        organization:
          nullable: true
          $ref: '#/components/schemas/Organization'
        job_title: { type: string }
        industries:
          type: array
          items: { type: string }
        competencies:
          type: array
          items: { type: string }
        tags:
          type: array
          items: { type: string }

        tie_strength: { $ref: '#/components/schemas/TieStrength' }

        trust_score:
          type: integer
          minimum: 1
          maximum: 10
          nullable: true
          description: May be redacted for assistants depending on privacy
        emotional_balance:
          type: number
          format: double
          nullable: true
          description: May be redacted for assistants depending on privacy

        how_can_help:
          type: array
          items: { type: string }
        how_i_can_help:
          type: array
          items: { type: string }

        met_context:
          type: string
          description: Where/how you met
        shared_notes:
          type: string
        private_notes:
          type: string
          description: Owner-only notes (may be null/redacted)

        last_interaction_at: { type: string, format: date-time, nullable: true }
        next_touch_at: { type: string, format: date-time, nullable: true }

        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      required:
        - id
        - workspace_id
        - visibility
        - display_name
        - tie_strength
        - created_at
        - updated_at
    ContactCreate:
      type: object
      properties:
        visibility: { $ref: '#/components/schemas/Visibility' }
        first_name: { type: string }
        last_name: { type: string }
        middle_name: { type: string }
        display_name: { type: string }
        photo_url: { type: string }
        emails:
          type: array
          items: { type: string, format: email }
        phones:
          type: array
          items: { type: string }
        messengers:
          type: object
          additionalProperties: { type: string }
        city: { type: string }
        timezone: { type: string }
        birthday: { type: string, format: date }
        organization_id: { $ref: '#/components/schemas/UUID' }
        job_title: { type: string }
        industries:
          type: array
          items: { type: string }
        competencies:
          type: array
          items: { type: string }
        tags:
          type: array
          items: { type: string }
        tie_strength: { $ref: '#/components/schemas/TieStrength' }
        trust_score:
          type: integer
          minimum: 1
          maximum: 10
          nullable: true
        emotional_balance:
          type: number
          format: double
          nullable: true
        how_can_help:
          type: array
          items: { type: string }
        how_i_can_help:
          type: array
          items: { type: string }
        met_context: { type: string }
        shared_notes: { type: string }
        private_notes: { type: string }
        next_touch_at: { type: string, format: date-time }
      required: [display_name, tie_strength]
    ContactUpdate:
      type: object
      properties:
        visibility: { $ref: '#/components/schemas/Visibility' }
        first_name: { type: string }
        last_name: { type: string }
        middle_name: { type: string }
        display_name: { type: string }
        photo_url: { type: string }
        emails:
          type: array
          items: { type: string, format: email }
        phones:
          type: array
          items: { type: string }
        messengers:
          type: object
          additionalProperties: { type: string }
        city: { type: string }
        timezone: { type: string }
        birthday: { type: string, format: date }
        organization_id: { $ref: '#/components/schemas/UUID' }
        job_title: { type: string }
        industries:
          type: array
          items: { type: string }
        competencies:
          type: array
          items: { type: string }
        tags:
          type: array
          items: { type: string }
        tie_strength: { $ref: '#/components/schemas/TieStrength' }
        trust_score:
          type: integer
          minimum: 1
          maximum: 10
          nullable: true
        emotional_balance:
          type: number
          format: double
          nullable: true
        how_can_help:
          type: array
          items: { type: string }
        how_i_can_help:
          type: array
          items: { type: string }
        met_context: { type: string }
        shared_notes: { type: string }
        private_notes: { type: string }
        next_touch_at: { type: string, format: date-time }
    Interaction:
      type: object
      properties:
        id: { $ref: '#/components/schemas/UUID' }
        workspace_id: { $ref: '#/components/schemas/UUID' }
        contact_id: { $ref: '#/components/schemas/UUID' }
        visibility: { $ref: '#/components/schemas/Visibility' }
        type: { $ref: '#/components/schemas/InteractionType' }
        channel: { $ref: '#/components/schemas/InteractionChannel' }
        occurred_at: { type: string, format: date-time }
        summary: { type: string }
        outcome: { type: string }
        next_action: { type: string }
        next_action_at: { type: string, format: date-time, nullable: true }
        created_by_user_id: { $ref: '#/components/schemas/UUID' }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      required: [id, workspace_id, contact_id, type, occurred_at, created_at, updated_at]
    InteractionCreate:
      type: object
      properties:
        visibility: { $ref: '#/components/schemas/Visibility' }
        type: { $ref: '#/components/schemas/InteractionType' }
        channel: { $ref: '#/components/schemas/InteractionChannel' }
        occurred_at: { type: string, format: date-time }
        summary: { type: string }
        outcome: { type: string }
        next_action: { type: string }
        next_action_at: { type: string, format: date-time }
      required: [type, occurred_at]
...
    Reminder:
      type: object
      properties:
        id: { $ref: '#/components/schemas/UUID' }
        workspace_id: { $ref: '#/components/schemas/UUID' }
        contact_id: { $ref: '#/components/schemas/UUID', nullable: true }
        project_id: { $ref: '#/components/schemas/UUID', nullable: true }
        type: { $ref: '#/components/schemas/ReminderType' }
        status: { $ref: '#/components/schemas/ReminderStatus' }
        title: { type: string }
        body: { type: string }
        due_at: { type: string, format: date-time }
        assigned_to_user_id: { $ref: '#/components/schemas/UUID', nullable: true }
        created_by_user_id: { $ref: '#/components/schemas/UUID' }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      required: [id, workspace_id, type, status, due_at, created_at, updated_at]
    ReminderCreate:
      type: object
      properties:
        contact_id: { $ref: '#/components/schemas/UUID' }
        project_id: { $ref: '#/components/schemas/UUID' }
        type: { $ref: '#/components/schemas/ReminderType' }
        title: { type: string }
        body: { type: string }
        due_at: { type: string, format: date-time }
        assigned_to_user_id: { $ref: '#/components/schemas/UUID' }
      required: [type, due_at]
    ReminderUpdate:
      type: object
      properties:
        status: { $ref: '#/components/schemas/ReminderStatus' }
        title: { type: string }
        body: { type: string }
        due_at: { type: string, format: date-time }
        assigned_to_user_id: { $ref: '#/components/schemas/UUID', nullable: true }
...
    ProjectCreate:
      type: object
      properties:
        visibility: { $ref: '#/components/schemas/Visibility' }
        name: { type: string }
        description: { type: string }
        status: { $ref: '#/components/schemas/ProjectStatus' }
        start_date: { type: string, format: date }
        end_date: { type: string, format: date }
        participant_contact_ids:
          type: array
          items: { $ref: '#/components/schemas/UUID' }
      required: [name]
...
    PaginatedContacts:
      type: object
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Contact' }
        next_cursor:
          type: string
          nullable: true
      required: [data]
...
    PaginatedInteractions:
      type: object
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Interaction' }
        next_cursor:
          type: string
          nullable: true
      required: [data]
...
    PaginatedReminders:
      type: object
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Reminder' }
        next_cursor:
          type: string
          nullable: true
      required: [data]
    PaginatedProjects:
      type: object
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Project' }
        next_cursor:
          type: string
          nullable: true
      required: [data]
...